Refactor + fix sso in LoginFragment
